implement new realloc and test rule in makefile

+++ tester/test.sh
@@ -22,18 +22,25 @@
 
 # ft_test "${1}" "${2}"
 
-make -C .. re
+REAL="real/"
+ME="me/"
 
-mkdir -p .tmp
-gcc -o ${1}_real real/${1}.c
-gcc -o ${2}_me me/${2}.c ../libft_malloc.so
-/usr/bin/time -v ./${1}_real 2> .tmp/out_real${1}
-/usr/bin/time -v ./${2}_me 2> .tmp/out_me${2}
-diff .tmp/out_real${1} .tmp/out_me${2} 2> .tmp/diff_out
-echo "diff :"
-cat .tmp/diff_out
-echo "page for real malloc :"
-cat .tmp/out_real${1} | grep Minor
-echo "page for my malloc :"
-cat .tmp/out_me${2} | grep Minor
-rm -rf .tmp ${1}_real ${2}_me
+if [ -f ${REAL}${1}.c ] && [ -f ${ME}${2}.c ];
+then
+    make -s -C .. re
+    mkdir -p .tmp
+    gcc -o ${1}_real ${REAL}${1}.c
+    gcc -o ${2}_me ${ME}${2}.c ../libft_malloc.so
+    /usr/bin/time -v ./${1}_real 2> .tmp/out_real${1}
+    /usr/bin/time -v ./${2}_me 2> .tmp/out_me${2}
+    echo "diff :"
+    diff .tmp/out_real${1} .tmp/out_me${2} 2> .tmp/diff_out
+    cat .tmp/diff_out
+    echo "page for real malloc :"
+    cat .tmp/out_real${1} | grep Minor
+    echo "page for my malloc :"
+    cat .tmp/out_me${2} | grep Minor
+    rm -rf .tmp ${1}_real ${2}_me
+else
+    echo File ${REAL}$1.c or ${ME}$2.c not found
+fi
